Added conversion to C
Added conversions to display the temperatures in degrees Celcius (just for the Pico S-C-Pro). All conversions are done to the display values, the JSON files and the data send to the Pico is still in F.
Some times there is one degree diference due to the decimals.
To use the Celcius mode an extra line is needed in the config.yaml

temperature: C

for Celcius or

temperature: F (or nothing)

For Fahrenheit